REFACTOR user modules to main modules
