+++ app/build.gradle
@@ -72,8 +72,8 @@ dependencies {
     implementation(project(":core:navigation"))
     implementation(project(":feature:home"))
     implementation(project(":feature:onboarding"))
-    implementation(project(":domain:user"))  // needed for di
-    implementation(project(":data:user"))  // needed for di
+    implementation(project(":domain:main"))  // needed for di
+    implementation(project(":data:main"))  // needed for di
     implementation(project(":core:network")) // needed for di
     implementation(libs.composeDestinations)
