+++ data/main/src/main/AndroidManifest.xml
