+++ data/main/src/main/kotlin/nl/q42/template/data/main/remote/UserApi.kt
@@ -1,0 +1,16 @@
+package nl.q42.template.data.main.remote
+
+import com.haroldadmin.cnradapter.NetworkResponse
+import nl.q42.template.actionresult.data.ApiErrorResponse
+import nl.q42.template.data.main.remote.model.UserDTO
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+internal interface UserApi {
+
+    /**
+     * note: [dummyEmailForResponse] is set because this test server mirrors the request as response.
+     */
+    @GET("get")
+    suspend fun getUsers(@Query("email") dummyEmailForResponse: String): NetworkResponse<UserDTO, ApiErrorResponse>
+}
