+++ data/main/src/main/kotlin/nl/q42/template/data/main/remote/UserRemoteDataSource.kt
@@ -1,0 +1,23 @@
+package nl.q42.template.data.main.remote
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import nl.q42.template.actionresult.data.mapToActionResult
+import nl.q42.template.actionresult.domain.ActionResult
+import nl.q42.template.actionresult.domain.map
+import nl.q42.template.data.main.local.model.UserEntity
+import nl.q42.template.data.main.mapper.mapToEntity
+import nl.q42.template.data.main.remote.model.UserDTO
+import javax.inject.Inject
+
+internal class UserRemoteDataSource @Inject constructor(
+    private val userApi: UserApi
+) {
+
+    suspend fun getUser(): ActionResult<UserEntity> = withContext(Dispatchers.IO) {
+        val apiActionResult = mapToActionResult {
+            userApi.getUsers("test@test.com")
+        }
+        apiActionResult.map(UserDTO::mapToEntity)
+    }
+}
