--- data/user/build.gradle
@@ -1,16 +1,0 @@
-plugins {
-    id "com.android.library"
-}
-
-apply from: "$rootDir/build.module.library.gradle"
-apply from: "$rootDir/build.dep.network.gradle"
-
-android {
-    namespace = "nl.q42.template.data.user"
-}
-
-dependencies {
-    implementation project(':domain:user')
-    implementation project(':core:network')
-    implementation project(':core:actionresult')
-}
