+++ domain/main/proguard-rules.pro
