+++ domain/main/src/main/AndroidManifest.xml
