+++ domain/main/src/main/kotlin/nl/q42/template/domain/main/model/User.kt
@@ -1,0 +1,6 @@
+package nl.q42.template.domain.main.model
+
+@JvmInline
+value class EmailAddress(val value: String)
+
+data class User(val email: EmailAddress)
