--- domain/user/src/main/kotlin/nl/q42/template/domain/user/model/User.kt
@@ -1,6 +1,0 @@
-package nl.q42.template.domain.user.model
-
-@JvmInline
-value class EmailAddress(val value: String)
-
-data class User(val email: EmailAddress)
