+++ feature/home/src/main/kotlin/nl/q42/template/home/main/presentation/HomeViewModel.kt
@@ -13,8 +13,8 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import nl.q42.template.actionresult.data.handleAction
-import nl.q42.template.domain.user.usecase.FetchUserUseCase
-import nl.q42.template.domain.user.usecase.GetUserFlowUseCase
+import nl.q42.template.domain.main.usecase.FetchUserUseCase
+import nl.q42.template.domain.main.usecase.GetUserFlowUseCase
 import nl.q42.template.feature.home.R
 import nl.q42.template.home.destinations.HomeSecondScreenDestination
 import nl.q42.template.navigation.AppGraphRoutes
